convert to uint32

+++ node/node_test.go
@@ -234,7 +234,7 @@ func TestCreateProposalBlock(t *testing.T) {
 	maxBytes := 16384
 	maxEvidence := 10
 	state.ConsensusParams.Block.MaxBytes = int64(maxBytes)
-	state.ConsensusParams.Evidence.MaxNumEvidence = int64(maxEvidence)
+	state.ConsensusParams.Evidence.MaxNumEvidence = uint32(maxEvidence)
 	proposerAddr, _ := state.Validators.GetByIndex(0)
 
 	// Make Mempool
